reading moench and raw files
