+++ core/include/aare/defs.hpp
@@ -34,7 +34,7 @@ struct xy {
 // using image_shape = std::array<ssize_t, 2>;
 using dynamic_shape = std::vector<ssize_t>;
 
-enum class DetectorType { Jungfrau, Eiger, Mythen3, Moench };
+enum class DetectorType { Jungfrau, Eiger, Mythen3, Moench,ChipTestBoard };
 
 enum class TimingMode {Auto, Trigger};
