+++ core/src/defs.cpp
@@ -10,13 +10,14 @@ template <> std::string toString(DetectorType type) {
         return "Mythen3";
     case DetectorType::Moench:
         return "Moench";
+    case DetectorType::ChipTestBoard:
+        return "ChipTestBoard";
     default:
         return "Unknown";
     }
 }
 
-
- template <> DetectorType StringTo(std::string name) {
+template <> DetectorType StringTo(std::string name) {
     if (name == "Jungfrau")
         return DetectorType::Jungfrau;
     else if (name == "Eiger")
@@ -25,18 +26,20 @@ template <> std::string toString(DetectorType type) {
         return DetectorType::Mythen3;
     else if (name == "Moench")
         return DetectorType::Moench;
+    else if (name == "ChipTestBoard")
+        return DetectorType::ChipTestBoard;
     else {
         auto msg = fmt::format("Could not decode dector from: \"{}\"", name);
         throw std::runtime_error(msg);
     }
 }
 
-template <> TimingMode StringTo(std::string mode){
+template <> TimingMode StringTo(std::string mode) {
     if (mode == "auto")
         return TimingMode::Auto;
-    else if(mode == "trigger")
+    else if (mode == "trigger")
         return TimingMode::Trigger;
-    else{
+    else {
         auto msg = fmt::format("Could not decode timing mode from: \"{}\"", mode);
         throw std::runtime_error(msg);
     }
