--- data/jungfrau_single_master_0.json
@@ -1,47 +1,0 @@
-{
-    "Version": 7.2,
-    "Timestamp": "Tue Feb 20 08:28:24 2024",
-    "Detector Type": "Jungfrau",
-    "Timing Mode": "auto",
-    "Geometry": {
-        "x": 1,
-        "y": 1
-    },
-    "Image Size in bytes": 1048576,
-    "Pixels": {
-        "x": 1024,
-        "y": 512
-    },
-    "Max Frames Per File": 3,
-    "Frame Discard Policy": "nodiscard",
-    "Frame Padding": 1,
-    "Scan Parameters": "[disabled]",
-    "Total Frames": 10,
-    "Receiver Roi": {
-        "xmin": 4294967295,
-        "xmax": 4294967295,
-        "ymin": 4294967295,
-        "ymax": 4294967295
-    },
-    "Exptime": "10us",
-    "Period": "1ms",
-    "Number of UDP Interfaces": 1,
-    "Number of rows": 512,
-    "Frames in File": 10,
-    "Frame Header Format": {
-        "Frame Number": "8 bytes",
-        "SubFrame Number/ExpLength": "4 bytes",
-        "Packet Number": "4 bytes",
-        "Bunch ID": "8 bytes",
-        "Timestamp": "8 bytes",
-        "Module Id": "2 bytes",
-        "Row": "2 bytes",
-        "Column": "2 bytes",
-        "Reserved": "2 bytes",
-        "Debug": "4 bytes",
-        "Round Robin Number": "2 bytes",
-        "Detector Type": "1 byte",
-        "Header Version": "1 byte",
-        "Packets Caught Mask": "64 bytes"
-    }
-}
