+++ data/moench/moench04_noise_200V_sto_both_100us_no_light_thresh_900_master_0.raw
@@ -1,0 +1,40 @@
+Version                    : 6.4
+TimeStamp                  : Wed Aug 31 09:08:49 2022
+
+Detector Type              : ChipTestBoard
+Timing Mode                : auto
+Geometry                   : [1, 1]
+Image Size                 : 360000 bytes
+Pixels                     : [96, 1]
+Max Frames Per File        : 20000
+Frame Discard Policy       : nodiscard
+Frame Padding              : 1
+Scan Parameters            : [disabled]
+Total Frames               : 100
+Exptime                    : 100us
+Period                     : 4ms
+Ten Giga                   : 1
+ADC Mask                   : 0xffffffff
+Analog Flag                : 1
+Analog Samples             : 5000
+Digital Flag               : 1
+Digital Samples            : 5000
+Dbit Offset                : 0
+Dbit Bitset                : 0
+Frames in File             : 100
+
+#Frame Header
+Frame Number               : 8 bytes
+SubFrame Number/ExpLength  : 4 bytes
+Packet Number              : 4 bytes
+Bunch ID                   : 8 bytes
+Timestamp                  : 8 bytes
+Module Id                  : 2 bytes
+Row                        : 2 bytes
+Column                     : 2 bytes
+Reserved                   : 2 bytes
+Debug                      : 4 bytes
+Round Robin Number         : 2 bytes
+Detector Type              : 1 byte
+Header Version             : 1 byte
+Packets Caught Mask        : 64 bytes
